feat: print order data

+++ OrderBook.cpp
@@ -9,7 +9,7 @@ std::priority_queue<Order> OrderBook::getAllOrders() {
 	return resting_orders;
 }
 
-void OrderBook::placeLimitOrder(std::string d, std::string t, float p, float s) {
+void OrderBook::placeLimitOrder(std::string d, std::string t, double p, double s) {
 	Order new_order = Order(d, t, p, s);
 	resting_orders.push(new_order);
 }
